Added box2D ipod project. Fixed compile error for GCC in box2D.
Added 64 bit colours, (RGBA u16 components) to see if the problem with the bad colours is a u8-specific issue on the ipod. It's probably something else, (Thought it might be colour conversion but then it should be broken on the PC too in that case) but added anyway.

Added a Debug scope timer (will compile out in release). Use in same way as normal one;
TLTime::Debug_ScopeTimer("Timer");

Added a TLCounter::Debug_Increment() which compiles out in release.

Corrected the maths and game ipod projects. The game one included itself in the sources. Maths I added new shape files (IIRC)

Added a custom contact filter for box2d to the physics graph. This should mean we do a few less body<->body collision tests. Can alter this further for more Tootle-specific collision-testing situations.
Added use of the GroupIndex in the body shapes so nodes set not to collide (eg. far-away cars) should no longer do collision tests.

Added an EnableCollision() function to the physics nodes which alters the flag but now does the box2d group-changing stuff

Added FORCE_COLOUR macro to TRenderTarget to hunt down this opengl render crash. Something wrong with the alternative colour buffer usage at the moment. Currently forcing to float colours which works fine. (But heavy on the GPU bus)

Scaled down the debug-text font in the screen's debug/stats render target when in landscape mode.

+++ Code/TootleMaths/TShapePolygon.cpp
@@ -73,10 +73,10 @@ Bool TLMaths::TShapePolygon2D::Debug_CheckIsConvex() const
 {
 	// Compute normals. Ensure the edges have non-zero length.
 	b2Vec2 m_normals[b2_maxPolygonVertices];
-	for (int32 i = 0; i < m_Outline.GetSize(); ++i)
+	for (int32 i = 0; i <(s32)m_Outline.GetSize(); ++i)
 	{
 		int32 i1 = i;
-		int32 i2 = i + 1 < m_Outline.GetSize() ? i + 1 : 0;
+		int32 i2 = i + 1 <(s32)m_Outline.GetSize() ? i + 1 : 0;
 		float2 edge2 = m_Outline[i2] - m_Outline[i1];
 		b2Vec2 edge( edge2.x, edge2.y );
 		b2Assert(edge.LengthSquared() > B2_FLT_EPSILON * B2_FLT_EPSILON);
@@ -85,9 +85,9 @@ Bool TLMaths::TShapePolygon2D::Debug_CheckIsConvex() const
 	}
 
 	// Ensure the polygon is convex.
-	for (int32 i = 0; i < m_Outline.GetSize(); ++i)
+	for (int32 i = 0; i <(s32)m_Outline.GetSize(); ++i)
 	{
-		for (int32 j = 0; j < m_Outline.GetSize(); ++j)
+		for (int32 j = 0; j <(s32)m_Outline.GetSize(); ++j)
 		{
 			// Don't check vertices on the current edge.
 			if (j == i || j == (i + 1) % m_Outline.GetSize())
@@ -122,10 +122,10 @@ Bool TLMaths::TShapePolygon2D::IsClockwise() const
 {
 	// Compute normals. Ensure the edges have non-zero length.
 	b2Vec2 m_normals[b2_maxPolygonVertices];
-	for (int32 i = 0; i < m_Outline.GetSize(); ++i)
+	for (int32 i = 0; i <(s32)m_Outline.GetSize(); ++i)
 	{
 		int32 i1 = i;
-		int32 i2 = i + 1 < m_Outline.GetSize() ? i + 1 : 0;
+		int32 i2 = i + 1 <(s32)m_Outline.GetSize() ? i + 1 : 0;
 		float2 edge2 = m_Outline[i2] - m_Outline[i1];
 		b2Vec2 edge( edge2.x, edge2.y );
 		b2Assert(edge.LengthSquared() > B2_FLT_EPSILON * B2_FLT_EPSILON);
@@ -134,7 +134,7 @@ Bool TLMaths::TShapePolygon2D::IsClockwise() const
 	}
 
 	// Ensure the polygon is counter-clockwise.
-	for (int32 i = 1; i < m_Outline.GetSize(); ++i)
+	for (int32 i = 1; i <(s32)m_Outline.GetSize(); ++i)
 	{
 		float32 cross = b2Cross(m_normals[i-1], m_normals[i]);
